bmstu/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-storages==1.14.6
boto3==1.42.61
botocore==1.42.61
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1

asgiref==3.11.1
jmespath==1.1.0
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3

ruff==0.6.9
pre-commit==3.8.0
